# Introduce __builtin_attr! to annotate 'inline'