--- a/integer.rb
+++ b/integer.rb
@@ -4,6 +4,7 @@ class Integer
   #
   # Returns +true+ if +num+ has a zero value.
   def zero?
+    Primitive.attr! 'inline'
     Primitive.cexpr! 'int_zero_p(self);'
   end
 end