https://bugs.ruby-lang.org/issues/15589

Based on https://github.com/ruby/ruby/pull/3226 which makes `Integer#zero?` safe for inlining, this PR introduces `__builtin_attr! 'inline'` to tell MJIT that builtin functions in an ISeq can be inlined. Note that [`__builtin*` DSL](https://bugs.ruby-lang.org/issues/16254) is private to Ruby core and it's not user-visible interface.

By just adding `__builtin_attr! 'inline'` on top of the `Integer#zero?` definition, `int.zero?` becomes as fast as `int == 0`.

```
$ benchmark-driver -v --rbenv 'before;before+JIT::before --jit;after+JIT::after --jit' benchmark/mjit_int_zero_p.yml --repeat-count=4
before: ruby 2.8.0dev (2020-06-14T22:01:15Z master 80bd3aa32b) [x86_64-linux]
before+JIT: ruby 2.8.0dev (2020-06-14T22:01:15Z master 80bd3aa32b) +JIT [x86_64-linux]
after+JIT: ruby 2.8.0dev (2020-06-14T22:48:16Z builtin-attr 01cbdaebcb) +JIT [x86_64-linux]
Calculating -------------------------------------
                         before  before+JIT   after+JIT
             0.zero?    44.391M     56.727M     70.439M i/s -     40.000M times in 0.901084s 0.705129s 0.567868s
             1.zero?    44.867M     56.989M     70.942M i/s -     40.000M times in 0.891531s 0.701895s 0.563843s
              0 == 0    68.324M     77.757M     77.322M i/s -     40.000M times in 0.585442s 0.514421s 0.517318s
              1 == 0    70.192M     77.372M     78.200M i/s -     40.000M times in 0.569866s 0.516984s 0.511508s

Comparison:
                          0.zero?
           after+JIT:  70438894.9 i/s
          before+JIT:  56727213.0 i/s - 1.24x  slower
              before:  44390986.6 i/s - 1.59x  slower

                          1.zero?
           after+JIT:  70941732.1 i/s
          before+JIT:  56988604.5 i/s - 1.24x  slower
              before:  44866660.1 i/s - 1.58x  slower

                           0 == 0
          before+JIT:  77757313.3 i/s
           after+JIT:  77321844.6 i/s - 1.01x  slower
              before:  68324430.6 i/s - 1.14x  slower

                           1 == 0
           after+JIT:  78200135.4 i/s
          before+JIT:  77371819.2 i/s - 1.01x  slower
              before:  70191899.3 i/s - 1.11x  slower
```

In this PR, there's no mechanism that verifies an ISeq with `__builtin_attr! 'inline'` is really safe for inlining. 

But we already know what to be checked https://bugs.ruby-lang.org/issues/16956, and we prove that kind of thing can be checked like `vm_check_canary` for the `leaf` attribute. So it'll be implemented once this gets merged, before starting to use the annotation in many places.